+++ app/src/main/java/com/vn/popcorn/beans/MovieItem.java
@@ -1,10 +1,9 @@
-package com.vn.popcorn;
+package com.vn.popcorn.beans;
 
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.provider.SyncStateContract;
 
 import org.json.JSONException;
 import org.json.JSONObject;
