+++ app/src/main/java/com/vn/popcorn/constants/Constants.java
@@ -1,3 +1,5 @@
+package com.vn.popcorn.constants;
+
 /**
  * Created by vn on 28/10/15.
  */
