+++ app/src/main/res/xml/pref_headers.xml
@@ -2,10 +2,10 @@
 
     <!-- These settings headers are only used on tablets. -->
 
-    <!--<header android:fragment="com.vn.popcorn.SettingsActivity$GeneralPreferenceFragment"-->
+    <!--<header android:fragment="com.vn.popcorn.activities.SettingsActivity$GeneralPreferenceFragment"-->
         <!--android:title="@string/pref_header_general" android:icon="@drawable/ic_info_black_24dp" />-->
 
-    <!--<header android:fragment="com.vn.popcorn.SettingsActivity$DataSyncPreferenceFragment"-->
+    <!--<header android:fragment="com.vn.popcorn.activities.SettingsActivity$DataSyncPreferenceFragment"-->
         <!--android:title="@string/pref_header_data_sync"  />-->
 
 </preference-headers>
